requirements.txt
fastapi
uvicorn
tensorflow==2.20.0
pillow
python-multipart
